requirement.txt
pymysql
websockets==13.1
requests
mysql-connector-python
schedule
flask
flask-cors
tradingview_ta
flask_cors
flask